move user model and its test function to model/user
